Move `StructuralIssues` to dedicated module
